test(mobile): split opfs-sqlite-driver.test.ts + apply prettier sweep
- opfs-sqlite-driver.test.ts (704 LOC) → 3 files by concern:
  * opfs-sqlite-driver-proxy-init.test.ts: init/prepare/all/exec/transactions (355 LOC)
  * opfs-sqlite-driver-proxy-errors.test.ts: close cleanup, worker error, init timeout, get, txn rollback (322 LOC)
  * opfs-sqlite-driver-robustness.test.ts: error kinds, per-call timeout, worker error handling, observability logs (384 LOC)
- Apply prettier --write to all in-scope files (CI was stricter than local format)

All in-scope files now ≤500 LOC. 142 test files, 1366 tests pass.
